[Docx Parser] Move style-parsing-specific code to a new module
